# Update brotli to v1.0.7, speed up MSVC-built versions