--- a/src/Native/AnyOS/brotli-version.txt
+++ b/src/Native/AnyOS/brotli-version.txt
@@ -1,2 +1,2 @@
-1.0.5
-https://github.com/google/brotli/releases/tag/v1.0.5
+1.0.7
+https://github.com/google/brotli/releases/tag/v1.0.7