--- a/src/Native/AnyOS/brotli/common/version.h
+++ b/src/Native/AnyOS/brotli/common/version.h
@@ -14,13 +14,13 @@
    BrotliEncoderVersion methods. */
 
 /* Semantic version, calculated as (MAJOR << 24) | (MINOR << 12) | PATCH */
-#define BROTLI_VERSION 0x1000005
+#define BROTLI_VERSION 0x1000007
 
 /* This macro is used by build system to produce Libtool-friendly soname. See
    https://www.gnu.org/software/libtool/manual/html_node/Libtool-versioning.html
  */
 
 /* ABI version, calculated as (CURRENT << 24) | (REVISION << 12) | AGE */
-#define BROTLI_ABI_VERSION 0x1005000
+#define BROTLI_ABI_VERSION 0x1007000
 
 #endif  /* BROTLI_COMMON_VERSION_H_ */