… src/Native/AnyOS/brotli/enc/backward_references_hq.c
… src/Native/AnyOS/brotli/enc/encode.c
… src/Native/AnyOS/brotli/enc/hash.h
… src/Native/AnyOS/brotli/enc/hash_composite_inc.h
… src/Native/AnyOS/brotli/enc/hash_longest_match_inc.h
… src/Native/AnyOS/brotli/enc/hash_rolling_inc.h
… src/Native/AnyOS/brotli/enc/hash_to_binary_tree_inc.h
… src/Native/AnyOS/brotli/enc/metablock.c
… src/Native/AnyOS/brotli/tools/brotli.c
… src/System.IO.Compression.Brotli/tests/Performance/BrotliPerfTests.cs