Brotli has been updated to v1.0.7 in the Google repo.  The first commit in this PR syncs up with the latest release tag (mostly ARM fixes and some code reformatting).

Additionally, there are 2 places in the encoder where GCC intrinsics for TZCNT and BSR are used.  I have Added the MSVC equivalents so that we don't have a performance penalty compared to GCC (and compatible) builds.  I submitted a [matching PR](https://github.com/google/brotli/pull/636) over in the Google repo, but the maintainers there have yet to respond.

I noticed a slight performance regression between 2.1 and the 3.0 previews, which had been updated to Brotli v1.0.5.  This change makes up that perf difference and more.

``` ini
BenchmarkDotNet=v0.11.3, OS=Windows 10.0.17134.556 (1803/April2018Update/Redstone4)
Intel Xeon CPU E3-1505M v6 3.00GHz, 1 CPU, 8 logical and 4 physical cores
Frequency=2929685 Hz, Resolution=341.3336 ns, Timer=TSC
.NET Core SDK=3.0.100-preview-010184
  [Host]        : .NET Core 2.1.7 (CoreCLR 4.6.27129.04, CoreFX 4.6.27129.04), 64bit RyuJIT
  2.1.7         : .NET Core 2.1.7 (CoreCLR 4.6.27129.04, CoreFX 4.6.27129.04), 64bit RyuJIT
  3.0-preview-2 : .NET Core 3.0.0-preview-27324-5 (CoreCLR 4.6.27322.0, CoreFX 4.7.19.7311), 64bit RyuJIT
  PR            : .NET Core 24973de7-9f09-4c96-a4b3-c4463fc81b91 (CoreCLR 4.6.27403.0, CoreFX 4.7.19.7311), 64bit RyuJIT

Jit=RyuJit  Platform=X64  Runtime=Core  

```
|          Method |           Job | Toolchain | Compressed Size |       Mean |     Error |    StdDev | Ratio | RatioSD |
|---------------- |-------------- |---------- |------ |-----------:|----------:|----------:|------:|--------:|
|  Alice29_q0_w12 |         2.1.7 |   Default | 78217 |   1.488 ms | 0.0237 ms | 0.0222 ms |  1.00 |    0.00 |
|  Alice29_q0_w12 | 3.0-preview-2 |   Default | 78217 |   1.499 ms | 0.0266 ms | 0.0249 ms |  1.01 |    0.00 |
|  Alice29_q0_w12 |            PR |   CoreRun | 78217 |   1.279 ms | 0.0149 ms | 0.0139 ms |  0.86 |    0.00 |
|                 |               |           |       |            |           |           |       |         |
|  Alice29_q5_w16 |         2.1.7 |   Default | 53089 |   7.308 ms | 0.0709 ms | 0.0663 ms |  1.00 |    0.00 |
|  Alice29_q5_w16 | 3.0-preview-2 |   Default | 53089 |   7.808 ms | 0.0862 ms | 0.0719 ms |  1.07 |    0.00 |
|  Alice29_q5_w16 |            PR |   CoreRun | 53089 |   6.061 ms | 0.0455 ms | 0.0404 ms |  0.83 |    0.01 |
|                 |               |           |       |            |           |           |       |         |
| Alice29_q11_w24 |         2.1.7 |   Default | 46487 | 275.866 ms | 2.6656 ms | 2.4934 ms |  1.00 |    0.00 |
| Alice29_q11_w24 | 3.0-preview-2 |   Default | 46487 | 286.640 ms | 2.5965 ms | 2.4288 ms |  1.04 |    0.00 |
| Alice29_q11_w24 |            PR |   CoreRun | 46487 | 260.619 ms | 2.2509 ms | 1.9954 ms |  0.94 |    0.00 |