@saucecontrol 2019-02-08
CI failures appear to be unrelated.

I also just noticed [brotli-version.txt](https://github.com/dotnet/corefx/blob/master/src/Native/AnyOS/brotli-version.txt) which appears to be redundant given the existence of [brotli/common/version.h](https://github.com/dotnet/corefx/blob/master/src/Native/AnyOS/brotli/common/version.h), which is updated in this PR.  Is it ok if I just remove brotli-version.txt rather than updating it?