@ahsonkhan 2019-02-09
> Is it ok if I just remove brotli-version.txt rather than updating it?

It was added based on feedback here: https://github.com/dotnet/corefx/pull/30751#issuecomment-401392909

I think we should continue to update it, if you don't mind.

cc @janvorli